Moved files to domain core and motion modules
