+++ src/domain/CMakeLists.txt
@@ -1,5 +1,2 @@
-target_sources(physics_backend
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/angle.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/vector.cpp
-)
+add_subdirectory(core)
+add_subdirectory(motion)
