+++ src/domain/core/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(physics_backend
+    PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/angle.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/vector.cpp
+)
