+++ src/domain/core/constants.hpp
