+++ src/domain/motion/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(physics_backend
+    PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/gravity.cpp
+)
