+++ test/domain/CMakeLists.txt
@@ -1,5 +1,1 @@
-target_sources(test_physics_backend 
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/test_angle.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/test_vector.cpp
-)
+add_subdirectory(core)
