+++ test/domain/core/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(test_physics_backend 
+    PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/test_angle.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/test_vector.cpp
+)
